refactor: NASA-99 refactor name constant anf file
